clause/clause_test.go
package clause

import (
	"reflect"
	"testing"
)

func testSelect(t *testing.T) {
	var clause Clause
	clause.Set(LIMIT, 3)
	clause.Set(SELECT, "User", []string{"*"})
	clause.Set(WHERE, "Name = ?", "Tom")
	clause.Set(ORDERBY, "Age ASC")
	sql, vars := clause.Build(SELECT, WHERE, ORDERBY, LIMIT)
	if sql != "SELECT * FROM User WHERE Name = ? ORDER BY Age ASC LIMIT ?" {
		t.Fatalf("failed to build SQL sql:%v", sql)
	}
	if !reflect.DeepEqual(vars, []interface{}{"Tom", 3}) {
		t.Fatalf("failed to build SQLVars sqlVars:%v", vars)
	}
	t.Log(sql, vars)
}

func TestClauses_Build(t *testing.T) {
	t.Run("select", func(t *testing.T) {
		testSelect(t)
	})
}
